Split Framework flex by type
