+++ internal/framework/flex/bool.go
@@ -1,0 +1,34 @@
+package flex
+
+import (
+	"context"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/hashicorp/terraform-plugin-framework/types"
+)
+
+// BoolFromFramework converts a Framework Bool value to a bool pointer.
+// A null Bool is converted to a nil bool pointer.
+func BoolFromFramework(_ context.Context, v types.Bool) *bool {
+	if v.IsNull() || v.IsUnknown() {
+		return nil
+	}
+
+	return aws.Bool(v.ValueBool())
+}
+
+// BoolToFramework converts a bool pointer to a Framework Bool value.
+// A nil bool pointer is converted to a null Bool.
+func BoolToFramework(_ context.Context, v *bool) types.Bool {
+	if v == nil {
+		return types.BoolNull()
+	}
+
+	return types.BoolValue(aws.ToBool(v))
+}
+
+// BoolToFrameworkLegacy converts a bool pointer to a Framework Bool value.
+// A nil bool pointer is converted to a false Bool.
+func BoolToFrameworkLegacy(_ context.Context, v *bool) types.Bool {
+	return types.BoolValue(aws.ToBool(v))
+}
