+++ internal/framework/flex/set.go
@@ -1,0 +1,112 @@
+package flex
+
+import (
+	"context"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/hashicorp/terraform-plugin-framework/attr"
+	"github.com/hashicorp/terraform-plugin-framework/types"
+)
+
+func ExpandFrameworkStringSet(ctx context.Context, set types.Set) []*string {
+	if set.IsNull() || set.IsUnknown() {
+		return nil
+	}
+
+	var vs []*string
+
+	if set.ElementsAs(ctx, &vs, false).HasError() {
+		return nil
+	}
+
+	return vs
+}
+
+func ExpandFrameworkStringValueSet(ctx context.Context, set types.Set) Set[string] {
+	if set.IsNull() || set.IsUnknown() {
+		return nil
+	}
+
+	var vs []string
+
+	if set.ElementsAs(ctx, &vs, false).HasError() {
+		return nil
+	}
+
+	return vs
+}
+
+func ExpandFrameworkStringValueMap(ctx context.Context, set types.Map) map[string]string {
+	if set.IsNull() || set.IsUnknown() {
+		return nil
+	}
+
+	var m map[string]string
+
+	if set.ElementsAs(ctx, &m, false).HasError() {
+		return nil
+	}
+
+	return m
+}
+
+// FlattenFrameworkStringSet converts a slice of string pointers to a framework Set value.
+//
+// A nil slice is converted to a null Set.
+// An empty slice is converted to a null Set.
+func FlattenFrameworkStringSet(_ context.Context, vs []*string) types.Set {
+	if len(vs) == 0 {
+		return types.SetNull(types.StringType)
+	}
+
+	elems := make([]attr.Value, len(vs))
+
+	for i, v := range vs {
+		elems[i] = types.StringValue(aws.ToString(v))
+	}
+
+	return types.SetValueMust(types.StringType, elems)
+}
+
+// FlattenFrameworkStringSetLegacy converts a slice of string pointers to a framework Set value.
+//
+// A nil slice is converted to an empty (non-null) Set.
+func FlattenFrameworkStringSetLegacy(_ context.Context, vs []*string) types.Set {
+	elems := make([]attr.Value, len(vs))
+
+	for i, v := range vs {
+		elems[i] = types.StringValue(aws.ToString(v))
+	}
+
+	return types.SetValueMust(types.StringType, elems)
+}
+
+// FlattenFrameworkStringValueSet converts a slice of string values to a framework Set value.
+//
+// A nil slice is converted to a null Set.
+// An empty slice is converted to a null Set.
+func FlattenFrameworkStringValueSet(_ context.Context, vs []string) types.Set {
+	if len(vs) == 0 {
+		return types.SetNull(types.StringType)
+	}
+
+	elems := make([]attr.Value, len(vs))
+
+	for i, v := range vs {
+		elems[i] = types.StringValue(v)
+	}
+
+	return types.SetValueMust(types.StringType, elems)
+}
+
+// FlattenFrameworkStringValueSetLegacy is the Plugin Framework variant of FlattenStringValueSet.
+// A nil slice is converted to an empty (non-null) Set.
+func FlattenFrameworkStringValueSetLegacy(_ context.Context, vs []string) types.Set {
+	elems := make([]attr.Value, len(vs))
+
+	for i, v := range vs {
+		elems[i] = types.StringValue(v)
+	}
+
+	return types.SetValueMust(types.StringType, elems)
+}
